hf-backend/requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
accelerate
ip-adapter
scikit-learn
numpy
Pillow
einops
python-multipart 